# deps: Optimize _add_incoming_needs()