This is a quick win, we can easily get rid of the quadratic runtime here.

-   In a repo with one node that has 2000 file items, the runtime of `bw test node-1` drops from ~8.4 seconds to ~4.7 seconds.
-   The “processing dependencies” stage of our Icinga node drops from ~32 seconds to ~21 seconds.